+++ src/Exceptions/ActionException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Bayfront\Bones\Exceptions;
+
+/*
+ * Thrown when there is an error constructing/retrieving an event.
+ */
+
+class ActionException extends BonesException
+{
+
+}